refactor(commands): extract appearance commands into appearance.rs
