+++ src-rust/crates/commands/src/lib.rs
@@ -226,6 +226,8 @@ mod teleport;
 pub use teleport::*;
 mod managed_agents;
 pub use managed_agents::*;
+mod appearance;
+pub use appearance::*;
 mod ui_settings;
 use ui_settings::*;
 
@@ -239,7 +241,6 @@ pub struct CompactCommand;
 pub struct CostCommand;
 pub struct ExitCommand;
 pub struct ModelCommand;
-pub struct ColorCommand;
 pub struct VersionCommand;
 pub struct ResumeCommand;
 pub struct StatusCommand;
@@ -264,10 +265,6 @@ pub struct RenameCommand;
 pub struct EffortCommand;
 pub struct SummaryCommand;
 pub struct CommitCommand;
-pub struct ThemeCommand;
-pub struct OutputStyleCommand;
-pub struct KeybindingsCommand;
-pub struct PrivacySettingsCommand;
 // Batch-1 new commands
 pub struct RemoteControlCommand;
 pub struct RemoteEnvCommand;
@@ -793,245 +790,6 @@ impl SlashCommand for ModelCommand {
                 new_config.provider = Some(provider.to_string());
             }
             CommandResult::ConfigChangeMessage(new_config, confirmation)
-        }
-    }
-}
-
-// ---- /color --------------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for ColorCommand {
-    fn name(&self) -> &str { "color" }
-    fn description(&self) -> &str { "Set or show the prompt bar color for this session" }
-    fn help(&self) -> &str {
-        "Usage: /color [<name|#RRGGBB|default>]\n\n\
-         Sets the accent color for the prompt bar in this session.\n\
-         Named colors: red, green, blue, yellow, cyan, magenta, white, orange, purple\n\
-         Hex codes:    #RGB or #RRGGBB\n\
-         Reset:        /color default\n\n\
-         The color is persisted to ~/.claurst/ui-settings.json and\n\
-         applied on the next REPL startup."
-    }
-
-    async fn execute(&self, args: &str, _ctx: &mut CommandContext) -> CommandResult {
-        let color = args.trim();
-        if color.is_empty() {
-            let current = load_ui_settings();
-            return CommandResult::Message(format!(
-                "Current prompt color: {}\n\
-                 Use /color <name|#RRGGBB|default> to change it.\n\n\
-                 Named colors: red, green, blue, yellow, cyan, magenta, white, orange, purple",
-                current.prompt_color.as_deref().unwrap_or("default"),
-            ));
-        }
-
-        let normalized = if color == "default" {
-            None
-        } else {
-            let known_colors = [
-                "red", "green", "blue", "yellow", "cyan", "magenta",
-                "white", "orange", "purple", "pink", "gray", "grey",
-            ];
-            let is_hex = color.starts_with('#') && (color.len() == 4 || color.len() == 7)
-                && color[1..].chars().all(|c| c.is_ascii_hexdigit());
-            if !is_hex && !known_colors.contains(&color.to_lowercase().as_str()) {
-                return CommandResult::Error(format!(
-                    "Unknown color '{}'. Use a color name (red, green, …) or a hex code (#RGB or #RRGGBB).",
-                    color
-                ));
-            }
-            Some(color.to_string())
-        };
-
-        match mutate_ui_settings(|s| s.prompt_color = normalized.clone()) {
-            Ok(_) => CommandResult::Message(format!(
-                "Prompt color set to {}.\n\
-                 Restart the REPL for the change to take effect.",
-                normalized.as_deref().unwrap_or("default")
-            )),
-            Err(e) => CommandResult::Error(format!("Failed to save color: {}", e)),
-        }
-    }
-}
-
-// ---- /theme --------------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for ThemeCommand {
-    fn name(&self) -> &str { "theme" }
-    fn description(&self) -> &str { "Show or change the current theme" }
-    fn help(&self) -> &str {
-        "Usage: /theme [default|dark|light]\n\
-         Without arguments, shows the active theme. With an argument, updates the theme for this and future sessions."
-    }
-
-    async fn execute(&self, args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let args = args.trim();
-        if args.is_empty() {
-            return CommandResult::Message(format!(
-                "Current theme: {:?}\nUse /theme <default|dark|light> to change it.",
-                ctx.config.theme
-            ));
-        }
-
-        let Some(theme) = parse_theme(args) else {
-            return CommandResult::Error(
-                "Theme must be one of: default, dark, light".to_string(),
-            );
-        };
-
-        let mut new_config = ctx.config.clone();
-        new_config.theme = theme.clone();
-        if let Err(err) = save_settings_mutation(|settings| settings.config.theme = theme.clone())
-        {
-            return CommandResult::Error(format!("Failed to save theme: {}", err));
-        }
-
-        CommandResult::ConfigChangeMessage(
-            new_config,
-            format!("Theme set to {}.", args.to_lowercase()),
-        )
-    }
-}
-
-// ---- /output-style -------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for OutputStyleCommand {
-    fn name(&self) -> &str { "output-style" }
-    fn description(&self) -> &str { "Show or switch the current output style" }
-    fn help(&self) -> &str {
-        "Usage: /output-style [style-name]\n\n\
-         With no argument: list available styles and show the current one.\n\
-         With a style name: switch to that style (persisted to settings).\n\n\
-         Built-in styles: default, verbose, concise\n\
-         Plugin-defined styles are listed automatically.\n\n\
-         Changes take effect on the next request."
-    }
-
-    async fn execute(&self, args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let arg = args.trim();
-        let valid_styles = available_output_style_names();
-        let current = current_output_style_name(&ctx.config);
-
-        if arg.is_empty() {
-            // List available styles
-            let mut lines = format!("Current output style: {}\n\nAvailable styles:\n", current);
-            for style in &valid_styles {
-                let marker = if style == current { " *" } else { "" };
-                lines.push_str(&format!("  {}{}\n", style, marker));
-            }
-            lines.push_str("\nUse /output-style <name> to switch.");
-            return CommandResult::Message(lines);
-        }
-
-        let normalized = arg.to_lowercase();
-        if !valid_styles.iter().any(|name| name == &normalized) {
-            return CommandResult::Error(format!(
-                "Unknown output style '{}'. Available styles: {}",
-                arg,
-                valid_styles.join(", ")
-            ));
-        }
-
-        let mut new_config = ctx.config.clone();
-        new_config.output_style = (normalized != "default").then(|| normalized.clone());
-        if let Err(err) = save_settings_mutation(|settings| {
-            settings.config.output_style =
-                (normalized != "default").then(|| normalized.clone());
-        }) {
-            return CommandResult::Error(format!("Failed to save configuration: {}", err));
-        }
-
-        CommandResult::ConfigChangeMessage(
-            new_config,
-            format!(
-                "Output style set to '{}'. Changes take effect on the next request.",
-                normalized
-            ),
-        )
-    }
-}
-
-// ---- /keybindings --------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for KeybindingsCommand {
-    fn name(&self) -> &str { "keybindings" }
-    fn description(&self) -> &str { "Create or open ~/.claurst/keybindings.json" }
-
-    async fn execute(&self, _args: &str, _ctx: &mut CommandContext) -> CommandResult {
-        let config_dir = Settings::config_dir();
-        let path = config_dir.join("keybindings.json");
-        let existed = path.exists();
-
-        if !existed {
-            if let Err(err) = std::fs::create_dir_all(&config_dir) {
-                return CommandResult::Error(format!(
-                    "Failed to create {}: {}",
-                    config_dir.display(),
-                    err
-                ));
-            }
-
-            let template = match generate_keybindings_template() {
-                Ok(template) => template,
-                Err(err) => {
-                    return CommandResult::Error(format!(
-                        "Failed to generate keybindings template: {}",
-                        err
-                    ))
-                }
-            };
-
-            if let Err(err) = std::fs::write(&path, template) {
-                return CommandResult::Error(format!(
-                    "Failed to write {}: {}",
-                    path.display(),
-                    err
-                ));
-            }
-        }
-
-        match open_with_system(&path.display().to_string()) {
-            Ok(_) => CommandResult::Message(if existed {
-                format!("Opened {} in your editor.", path.display())
-            } else {
-                format!(
-                    "Created {} with a template and opened it in your editor.",
-                    path.display()
-                )
-            }),
-            Err(err) => CommandResult::Message(if existed {
-                format!(
-                    "Opened {}. Could not launch an editor automatically: {}",
-                    path.display(),
-                    err
-                )
-            } else {
-                format!(
-                    "Created {} with a template. Could not launch an editor automatically: {}",
-                    path.display(),
-                    err
-                )
-            }),
-        }
-    }
-}
-
-// ---- /privacy-settings ---------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for PrivacySettingsCommand {
-    fn name(&self) -> &str { "privacy-settings" }
-    fn description(&self) -> &str { "Open Claurst privacy settings" }
-
-    async fn execute(&self, _args: &str, _ctx: &mut CommandContext) -> CommandResult {
-        let url = "https://claude.ai/settings/data-privacy-controls";
-        let fallback = format!("Review and manage your privacy settings at {}", url);
-        match open_with_system(url) {
-            Ok(_) => CommandResult::Message(format!("Opened privacy settings: {}", url)),
-            Err(_) => CommandResult::Message(fallback),
         }
     }
 }
